Integrated the styling layout components to the map explorer controls
The main styling work using tailwindcss and heroicons has been
integrated into the layout via a nested Route thing. There are still a
few issues that seem to be due to the header not being friends with full
screen maps. Trying to address that now.

Coexistence between bootstrap and tailwind css is hard work as they
share the same names for utility classes and bootstrap does not offer an
obvious prefixing solution. Happily, leaflet styling is somewhat more
polite. All this to explain why we now have sass in our pipeline. I hope
we can remove it as we shouldn't need to do all the weird css stuff
eventually.