Integrated the site header and menu properly
Converted a bit more to tailwind and pushed down the shims for resetting
back to bootstrap down a layer - including porting the spinner to favour
the tailwindcss extensions added to mimic the bootstrap equivalent. Made
sure that the header integrates tolerably with the full screen map and
list views and fixed a couple of layout display issues on the hamburger.