Create a build per subdomain which includes a custom vite param for location (see notes)
Moving away from the domain-name lookup for resolving the location parameters. This allows us to move to a single domain without too much hassle.

Of course now we have to build several times instead of once but it's not super slow and you can do both steps separately if you want.