Big old rewrite of existing hooks and app model code (see notes)
This code was getting ridiculously unmaintainable so I've added a new set of hooks and models
that we can swap the components to use. Then we can rename them to more long-term names.

src/hooks/api.js

Has all the hooks for getting to the application data.
Added in a somewhat problematic abstraction ('datasources') to allow per-route datasources without all the components
needing to know what kind of datasource it is.

The upshot being that the map does not care if it is a filtered gig list or an individual gig. It just gets whatever shaped data it wants from whatever the current datasource is.

Using the 'handle' prop in react router to associate datasources to routes and then decide based on the datasource what to load and how to load it. We use the 'expectedType' to make sure that useGigList() only returns a valid result when we are on the gig list page.

Mainly achieving this via the useSWR conditional-request type property - if there is no requestkey, no request is made.

Honestly it's a bit dodgy and I am not sure it really is worth it. But it is tested and works right now.

src/model_v2.js

This is more or less completely different to src/model.js. Basically a bunch of in memory manipulation of data - converting raw API responses to useful data, that kind of thing. It
converts the api response in two stages - once to map the incoming data to useful structures like tag objects not string lists, then a second time to do the in memory filtering of tags and venues etc.
The idea being that we only need to do the first transform when a server side filter changes (and we doa new request), and the second transform (applyFilters) when an in mem filter changes. Much optimisation has not been done and is probably possible.

src/hooks/searchParams_v2.js

Largely inherited from the mess in the current src/hooks/filters.js code - it really needs a lot of simplification as it's just some object transforms and react router hook access.