Updating Map component to use new hooks (see notes)
This was the main focus for changing the api hooks, so it's got a bunch of changes in it.

Firstly removed a lot of the in memory processing of venues and gigs - that's all coped with by the data that is returned from the location-specific hooks (useCurrentLocationSettings and useMapVenues)

useCurrentLocationSettings and useMapVenues provide valid responses both when a single gig is being viewed, and when the gig list is being viewed.

useCurrentLocationSettings has been made a little bit more complicated to attempt to only change the location when the navigation suggests it is worth doing (like you click on a gig, or change the location filter).

Hopefully this means that the user's scrolling and panning is not disrupted too much as they use the filters. Not perfect but maybe a good start.

The gig details view now uses the venue location to show the correct pin. That's a big change and one that is a bit of an issue as we mainly obscure half of the map with the gig list - so the centre of the map is not really centred.
All our existing location specific map coordinates are offset a bit to cope with this.