Configuring leaflet to be an external dependency (see notes)
mapping itself as a global "L" which is apparently OK for react-leaflet to pick up.

vite has been configured to separate these and mention them all in the manifest so the rails loader
code on the rails app can deal with the extra css and js files.

have used a CDN for the index.html file as it's only for dev - it would be good to use the local assets here but
I couldn't be bothered to do that right now!